src/stress_tests/requirements/base.txt
# pulls in lots of libraries...
locustio==0.7.2